// CoreConstants.d.ts
export declare const CoreConstants: {
    PollingIntervalInMs: number;
    Headers: {
        Bearer: string;
        SdkVersionHeaderName: string;
        SdkVersionHeaderValueFormatString: string;
        FormUrlEncodedContentType: string;
        ThrowSiteHeaderName: string;
    };
    Serialization: {
        ODataType: string;
    };
};